removed table settings view
