+++ css/dataCanvas.css
@@ -1,10 +1,34 @@
 .data-canvas-view .data-canvas-item-view {
+    position: absolute;
+    border: 1px solid #008CBA;
+    width: 200px;
+}
+
+.data-canvas-view .data-canvas-item-view .title {
     background: #008CBA;
     color: white;
     cursor: default;
-    width: 150px;
-    padding-left: 25px;
-    position: absolute;
+    padding: 6px 25px;
+}
+
+.data-canvas-view .data-canvas-item-view ul {
+    margin: 0;
+    max-height: 200px;
+    overflow: auto;
+    background: white;
+}
+.data-canvas-view .data-canvas-item-view li label {
+    padding: 0 10px;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    color: #008CBA;
+    white-space: nowrap;
+}
+
+.data-canvas-view .data-canvas-item-view li input {
+    margin: 0 5px 0 0;
+    position: relative;
+    top: 2px;
 }
 
 .canvas-items-holder {
