+++ css/main.css
@@ -17,7 +17,6 @@
 @import "dataCanvas.css";
 @import "dataCanvas.css";
 @import "axis.css";
-@import "tableSettings.css";
 @import "tables.css";
 @import "dialog.css";
 @import "domains.css";
