--- css/tableSettings.css
@@ -1,36 +1,0 @@
-.table-settings-view {
-    border-left: 1px solid #007095;
-}
-
-.table-settings-view .title,
-.table-settings-view .subtitle {
-    overflow: hidden;
-    text-overflow: ellipsis;
-    padding-left: 20px;
-    padding-right: 20px;
-}
-
-.table-settings-view .subtitle {
-    background: #008CBA;
-    color: white;
-}
-
-.table-settings-view ul {
-    overflow: auto;
-}
-
-.table-settings-view ul li {
-    white-space: nowrap;
-}
-
-.table-settings-view li label{
-    color: #008CBA;
-    overflow: hidden;
-    text-overflow: ellipsis;
-}
-
-.table-settings-view li input {
-    margin: 0 5px 0 0;
-    position: relative;
-    top: 2px;
-}
