+++ src/template/DataCanvasViewTemplate.html
@@ -1,2 +1,1 @@
-<div class="canvas-items-holder column medium-10"></div>
-<div class="table-settings-holder column medium-2"></div>
+<div class="canvas-items-holder"></div>
